+++ db/migrate/20210812152307_create_month_fecals.rb
@@ -2,9 +2,9 @@ class CreateMonthFecals < ActiveRecord::Migration
   def change
     create_table :month_fecals do |t|
     
-      t.float :up_std,  null: false, default: Setting.systems.default_num 
+      t.integer :up_std,  null: false, default: Setting.systems.default_num 
     
-      t.float :end_std,  null: false, default: Setting.systems.default_num 
+      t.integer :end_std,  null: false, default: Setting.systems.default_num 
     
       t.float :yoy,  null: false, default: Setting.systems.default_num
